rewrite teaser and optimize other stories as well
