+++ packages/components/src/components/button/button.test.stories.ts
@@ -19,16 +19,12 @@ export default {
   title: 'Components/sd-button/Test',
   component: 'sd-button',
   tags: ['!autodocs'],
-  parameters: {
-    ...parameters,
-    controls: { include: [] },
-    design: {
-      type: 'figma',
-      url: 'https://www.figma.com/file/fPGhgNZv98U4H69Gu2tlWi/Button?type=design&node-id=13-18&t=jDLqFEdY7ZlOJurc-4'
-    }
-  },
   args: overrideArgs({ type: 'slot', name: 'default', value: 'Default' }),
   argTypes,
+  parameters: {
+    ...parameters,
+    controls: { include: [] }
+  },
   decorators: [withActions] as any
 };
