+++ packages/components/src/templates/quote.stories.ts
@@ -1,10 +1,9 @@
-import '../../solid-components';
+import '../solid-components';
 import { html } from 'lit-html';
 
 export default {
   tags: ['!dev'],
   title: 'Templates/Quote',
-  component: 'Quote',
   parameters: {
     chromatic: { disableSnapshot: true },
     design: {
